backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
attrs==25.4.0
bcrypt==3.2.2
billiard==4.2.4
celery==5.3.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.4
dnspython==2.8.0
ecdsa==0.19.1
elastic-transport==9.2.1
elasticsearch==9.3.0
email-validator==2.1.0
exceptiongroup==1.3.1
fastapi==0.104.1
flower==2.0.1
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
humanize==4.15.0
idna==3.11
iniconfig==2.3.0
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.9
pyasn1==0.6.2
pycparser==3.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
PyJWT==2.11.0
pytest==7.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.4.0
tornado==6.5.4
twilio==9.0.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.24.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.5.3
websockets==16.0
yarl==1.22.0
